# Change bsearch's `/ 2` to `>> 1` for faster performance