@asterite 2019-03-09
The reason is that division checks for zero and might raise. That makes it slower. I think there's also unsafe_div. But this is fine.