@Fryguy 2019-03-09
I had originally tried unsafe_div and it was slower than `>> 1`.  I had also tried unsafe_shr, but there was no change over `>> 1`