@asterite 2019-03-09
Actually /2 shouldn't be slower because it's a hardcoded value. This is strange...