@Fryguy 2019-03-09
I was actually surprised that LLVM didn't have a built-in optimization for `/ 2`