@Fryguy 2019-03-09
FWIW:

/ 2
```
Array#bsearch   6.83M (146.38ns) (± 2.92%)  0 B/op        fastest
Range#bsearch   6.11M ( 163.7ns) (± 2.31%)  0 B/op   1.12× slower
```

\>> 1
```
Array#bsearch   13.1M ( 76.35ns) (± 3.12%)  0 B/op        fastest
Range#bsearch   9.88M (101.23ns) (± 3.00%)  0 B/op   1.33× slower
```

unsafe_div(2)
```
Array#bsearch   9.09M (109.97ns) (± 2.65%)  0 B/op        fastest
Range#bsearch   7.35M (136.13ns) (± 3.28%)  0 B/op   1.24× slower
```

unsafe_shr(1)
```
Array#bsearch  13.06M ( 76.54ns) (± 3.66%)  0 B/op        fastest
Range#bsearch   9.92M (100.84ns) (± 3.46%)  0 B/op   1.32× slower
```