@konovod 2019-03-09
Signed division could be slower because it rounds up to zero (at least as [LLVM docs](https://llvm.org/docs/LangRef.html#sdiv-instruction) say), and shift rounds to negative infinity. Benchmarking confirms that unsigned division has same speed and signed is 1.23 times slower. 

On the other hand looks like signed division in Crystal always rounds to negative infinity - i have no idea how to explain it.