@Fryguy 2019-03-09
Just tried being more explicit by doing an unsigned div with `/ 2_u32` and `/2_u8`.  No change from regular `/ 2`.  Updated my post above.