@konovod 2019-03-09
The left side is still signed. Can't currently check myself, but I'm sure that if you will convert left side to UInt32 (unsafely) and then do unsafe_div, performance of bsearch will match the one with `>> 1`.