@Fryguy 2019-03-09
@konovod Doesn't seem to have worked, but it's possible I did it wrong:

```
Array#bsearch   6.91M (144.62ns) (± 2.82%)  0 B/op        fastest
Range#bsearch   6.16M (162.37ns) (± 3.25%)  0 B/op   1.12× slower
```

Is this the "right" way to do it?

```diff
-      mid = from + ((to - from) >> 1)
+      mid = from + ((to - from).unsafe_as(Int32) / 2_u32)
```