@konovod 2019-03-09
unsafe_as(UInt32)