@Fryguy 2019-03-09
LOL oops..of course :smile:

So, yes, using `.unsafe_as(Int32).unsafe_div(2_u32)` has similar performance to `>> 1`

```
Array#bsearch  13.19M ( 75.81ns) (± 3.50%)  0 B/op        fastest
Range#bsearch   9.92M (100.76ns) (± 3.81%)  0 B/op   1.33× slower
```