@konovod 2019-03-09
I think that in this (and similar) places, `>> 1` is easier to understand and looks better than converting to unsigned.
For a general case (unsigned division by, say, 3 is faster then signed too), maybe there should be something like `unsafe_positive_div` but this looks like a low priority and debatable thing.